refactor: extract rag_text_processing.py from rag_core (removes late-import cycle) (#545)
rag_core.py에서 텍스트 처리 primitives를 rag_text_processing.py로 추출.
이 모듈은 leaf — stdlib + korean_lexicon만 의존하여 순환 import 없음.

이동된 심볼 (14개):
- tokenize, sentence_split, split_long_text_unit
- normalize_entity, normalize_metadata_token, normalize_section_path
- ordered_unique, coerce_string_list, coerce_alias_values
- compact_metadata_text
- ENTITY_RE, TOKEN_RE, SENTENCE_RE, QUERY_TYPE_TOP_K_DEFAULTS

late-import → module-level import 전환:
- rag_verifier: normalize_metadata_token, ordered_unique
- rag_answer: sentence_split, tokenize, ordered_unique, compact_metadata_text
- rag_retrieval: tokenize
- rag_query: ENTITY_RE, QUERY_TYPE_TOP_K_DEFAULTS, coerce_string_list,
  compact_metadata_text, normalize_entity, normalize_metadata_token,
  ordered_unique, tokenize

rag_core backward-compat re-export 유지 — 기존 테스트/스크립트 unbroken.

Closes #545
